Make copystatic command more graceful on missing directories.
git-svn-id: http://trac.edgewall.org/intertrac/log:/trunk@6730 af82e41b-90c4-0310-8c96-b1721e28e2e2

+++ trac/admin/console.py
@@ -1173,9 +1173,9 @@ Congratulations!
             for key, root in paths:
                 source = os.path.normpath(root)
                 print '   ', source
-                dest = os.path.join(target, key)
-                copytree(source, dest)
-        
+                if os.path.exists(source):
+                    dest = os.path.join(target, key)
+                    copytree(source, dest)
 
 
 class TracAdminHelpMacro(WikiMacroBase):
